+++ src/main/java/com/eunsunzzang/posting/member/model/EmailAuthStatus.java
@@ -1,4 +1,4 @@
-package com.eunsunzzang.posting.member;
+package com.eunsunzzang.posting.member.model;
 
 public enum EmailAuthStatus {
     VERIFICATION_ING,
