+++ src/main/java/com/eunsunzzang/posting/member/model/MemberRole.java
@@ -1,4 +1,4 @@
-package com.eunsunzzang.posting.member;
+package com.eunsunzzang.posting.member.model;
 
 public enum MemberRole {
     ROLE_USER,
