+++ src/main/java/com/eunsunzzang/posting/member/model/dto/EmailAuthDto.java
@@ -1,7 +1,6 @@
-package com.eunsunzzang.posting.member.dto;
+package com.eunsunzzang.posting.member.model.dto;
 
 public record EmailAuthDto(
         String email,
-
         String key) {
 }
