+++ src/main/java/com/eunsunzzang/posting/member/model/dto/LoginDto.java
@@ -1,0 +1,6 @@
+package com.eunsunzzang.posting.member.model.dto;
+
+public record LoginDto(
+        String email,
+        String password) {
+}
