+++ src/main/java/com/eunsunzzang/posting/member/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package com.eunsunzzang.posting.member.repository;
 
-import com.eunsunzzang.posting.member.Member;
+import com.eunsunzzang.posting.member.model.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
