+++ src/test/java/com/eunsunzzang/posting/member/service/MemberServiceTest.java
@@ -1,11 +1,10 @@
 package com.eunsunzzang.posting.member.service;
 
 import com.eunsunzzang.posting.error.errorcode.AuthErrorCode;
-import com.eunsunzzang.posting.member.Member;
-import com.eunsunzzang.posting.member.MemberRole;
-import com.eunsunzzang.posting.member.dto.MemberSignUpDto;
+import com.eunsunzzang.posting.member.model.entity.Member;
+import com.eunsunzzang.posting.member.model.MemberRole;
+import com.eunsunzzang.posting.member.model.dto.MemberSignUpDto;
 import com.eunsunzzang.posting.member.repository.MemberRepository;
-import com.eunsunzzang.posting.member.service.MemberService;
 import jakarta.persistence.EntityManager;
 import jakarta.transaction.Transactional;
 import org.junit.jupiter.api.DisplayName;
